Added the populate option type string or string[] to the methods delete() and deleteAll()

+++ src/shared/intraestructure/MongoRepository.ts
@@ -42,17 +42,39 @@ export class MongoRepository<T extends Document> implements IRepository<T> {
         return await this.model.findOneAndUpdate(filter, data, { new: true }).exec();
     }
 
-    async delete(filter: Record<string, any>): Promise<T | null> {
-        return await this.model.findOneAndDelete(filter).exec();
+    async delete(filter: Record<string, any>, populate?: string[] | string): Promise<T | null> {
+        if (!populate) return await this.model.findOneAndDelete(filter).exec();
+        
+        if (!Array.isArray(populate)) return await this.model.findOneAndDelete(filter).populate(populate).exec();
+
+        let query = this.model.findOne(filter);
+        
+        populate.forEach((field) => {
+            query = query.populate(field);
+        });
+
+        return await query.exec();
     }
 
-    async deleteAll (filter: Record<string, any>): Promise <T[]> {
-        const documents = await this.model.find(filter).exec();
+    async deleteAll (filter: Record<string, any>, populate?: string[] | string): Promise <T[]> {
+        let query = this.model.find(filter);
 
+        if (populate) {
+            if (Array.isArray(populate)) {
+                populate.forEach((field) => {
+                    query = query.populate(field);
+                });
+            } else {
+                query = query.populate(populate);
+            }
+        }
+    
+        const documents = await query.exec();
+    
         if (documents.length > 0) {
             await this.model.deleteMany(filter).exec();
         }
-
-        return documents
+    
+        return documents;
     }
 }
